Set up React Router for front end
