+++ The-Coding-Sharks-UI/src/components/home.jsx
@@ -1,0 +1,21 @@
+import { Login } from './Login';
+import { Register } from './Register';
+import { useState } from "react";
+
+function Home() {
+  const [currentForm, setCurrentForm] = useState('login');
+
+  const toggleForm = (formName) => {
+    setCurrentForm(formName);
+  }
+
+  return (
+    <div className="App">
+        {
+          currentForm === "login" ? <Login onFormSwitch={toggleForm} /> : <Register onFormSwitch={toggleForm} />
+        }
+      </div>
+  )
+};
+
+export default Home;
